Schema updated to include part/measure/directions/textType...
Issue: I'm currently assuming that:
If align has one of the values 'left', 'right' or 'centre', the text will be aligned  to the event whose location is given by the location attribute.
If align has one of the values 'barlineLeft', 'barlineRight' or 'barlineCentre', the text will be aligned to the barline at the beginning of the measure, and the location attribute will be ignored.
Is that okay?